♻️ refactor: utils 함수 성격별 분류 및 주석 추가 - 함수 목적에 따라 utils를 성격별로 재구성 - 각 함수의 용도를 명확히 하기 위해 주석 추가
